Movie views renamed and templated added
