+++ apps/movie/views/core.py
